Move external libraries into arm9 subregion
This region is probably going to get much larger in the near future.
Since arm9.yml is already huge as is, and there's a natural boundary,
creating a subregion makes sense.
